requirements.txt
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docker-pycreds==0.4.0
filelock==3.18.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
pillow==11.1.0
platformdirs==4.3.7
protobuf==5.29.4
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.1
sentry-sdk==2.23.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.8
